Add files via upload
Problem Statement: 
In wildlife safaris, when the visibility is low, animals become difficult to spot.

My Solution;
Using Arduino Uno and components like Ultrasonic Sensor,Bluetooth and RGB, i have created a project such that when the dist between animal and the sensor is greater than 50cm, the RGB gives a green glow at a slow speed.
Similarly if distance is between 25 and 50 cm, blue light flickers at moderate speed and lastly, if distance is less than 25cm, a red light is seen blinking very fast.
So, this can be useful to tourists and guides during the night time and when the visibility decreases. 
This can be made on a larger scale and can be installed in the lodges located right between the forests.